Move fixtures out of plugin module to keep coverage
